Fix implementation of habitatFeatures [#177070719] (#65)
In addition to fixing the inappropriate persistence of habitat features across containers, we also take this opportunity to eliminate some LeafPack-specific implementation details that had crept into the implementation of the `useModelState()` hook. Specifically, rather than accepting a map of initial output states (which requires that `useModelState()` know something about the containers required by LeafPack), the `initialInputState()` and `initialOutputState()` properties are now functions which return the required states on demand.

Finally, there is a unit test of the `useLeafPackModelState()` hook which tests that habitat features are preserved when they should be and not when they shouldn't be as well as that initial fish configurations can differ, etc.