Update @concord-consortium/react-components; add <ControlPanel/> unit test (#71)
Previously, components that used components from `@concord-components/react-components` weren't unit-testable due to the fact that `@concord-components/react-components` was published as es6 modules but jest doesn't yet support es6 modules. Therefore, `@concord-components/react-components` was updated to publish as a commonjs module, and this PR updates to the newer version and adds a unit test for the `<ControlPanel/>` component.