Tweaks to <SimulationAnimation> component (#72)
Well, _that_ was more of an odyssey than I expected. The original issue reported by @scytacki was that in its steady-state, when viewed in a browser with caching disabled (not an expected end-user situation), leaf-pack continuously loads the images used for animation (i.e. the network panel continues to show activity for each animation frame). It was hypothesized that this might lead to performance issues as well, e.g. higher CPU usage.

The relatively simple fix which was straightforward to implement was to render all frames into the DOM and use CSS `visibility` to control which frame is visible rather than rendering only a single frame at a time into the DOM as we had been before. In a parallel universe that makes sense this would have been the end of the story. In _this_ universe, however, it was just my entree into the rabbit hole.

With the fish animations, unlike any of the other animations, the animation occurs within an `opacity` transition, because the fish fade in gradually while they're animating/swimming. This `opacity` transition was handled with ReactTransitionGroup's `<TransitionGroup>` and `<CSSTransition>` components, which under the hood make use of a CSS transition on `opacity`. After making the change to control frame visibility through CSS `visibility` rather than through presence/absence in the DOM, there was obvious and distracting flashing that occurred. Looking closely it was apparent that the images were flickering between the properly translucent state consistent with their `opacity` and their final state somewhat randomly. After beating my head against this for several hours trying to figure out what I was doing wrong, I eventually came to the conclusion that browsers apparently can't handle simultaneously changing CSS properties while transitioning another CSS property (e.g. `opacity`). I tried using position (`left`, `right`) to control visibility instead of `visibility` but the effect was the same. This [issue](https://stackoverflow.com/questions/24465619/opacity-change-during-a-transition-flickers-in-safari) was raised on Stack Overflow for Safari years ago (by none other than React's Dan Abramov) but now appears to affect Chrome and Firefox on Mac and Windows as well.

Eventually, it occurred to me that I could eliminate the `<TransitionGroup>` and `<CSSTransition>` components and just handle the `opacity` inside the animation component itself with a timer. Now the `opacity` is handled by the `<SimulationAnimation>` component itself rather than relying on a CSS transition applied by the parent in the case of the fish animation. In theory this could lead to a user-visible difference in the case of slow animations with longer durations on each frame, in which case the previous implementation would continue to change the `opacity` on each frame while the new implementation only changes the opacity on frame transitions, but in most real-world scenarios this difference is unlikely to be noticeable. Phew!